requirements.txt
langchain>=0.3.0
langgraph>=0.1.20
langchain-community>=0.3.0
langchain-openai>=0.2.0
pillow>=10.3.0
httpx>=0.27.0
tavily-python>=0.3.5
python-whois>=0.9.4
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.4
